+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Preparations to routing/link building code for future porting to Symfony controllers !2738 @\_fridtjof_
 - Moved the database connection to Doctrine DBAL !2801 @\_fridtjof_
 - Refactored the list of editable content to vue !2862 @alex.simm
+- Modernized ContentControl to a Symfony controller !2876 @\_fridtjof_
 
 ## Dev/Test/CI stuff
 - removed CHANGELOG.md merge=union !2866 @chriswalg